fix: PixivDeleteCommand

+++ helper/src/main/kotlin/xyz/cssxsh/mirai/plugin/command/PixivDeleteCommand.kt
@@ -5,6 +5,7 @@ import net.mamoe.mirai.console.command.CommandSender
 import net.mamoe.mirai.utils.*
 import xyz.cssxsh.mirai.plugin.*
 import xyz.cssxsh.mirai.plugin.data.*
+import xyz.cssxsh.mirai.plugin.model.*
 import java.time.OffsetDateTime
 
 object PixivDeleteCommand : CompositeCommand(
@@ -14,7 +15,7 @@ object PixivDeleteCommand : CompositeCommand(
     overrideContext = PixivCommandArgumentContext
 ) {
 
-    private fun deleteImage(pid: Long): Boolean {
+    private fun ArtWorkInfo.delete(): Boolean {
         return PixivHelperSettings.imagesFolder(pid).listFiles { file ->
             file.isFile && file.extension != "json"
         }?.all { it.delete() } ?: true
@@ -22,26 +23,25 @@ object PixivDeleteCommand : CompositeCommand(
 
     @SubCommand
     @Description("删除指定作品")
-    suspend fun CommandSender.artwork(pid: Long) {
+    suspend fun CommandSender.artwork(pid: Long, record: Boolean = false) {
         logger.info { "作品(${pid})信息将从缓存移除" }
-        useMappers {
-            it.artwork.deleteByPid(pid = pid, comment = "command delete artwork in ${OffsetDateTime.now()}")
+        if (record) useMappers { mappers ->
+            mappers.artwork.deleteByPid(pid = pid, comment = "command delete artwork in ${OffsetDateTime.now()}")
         }
-        logger.info { "作品(${pid})信息将从缓存移除" }
-        sendMessage("作品(${pid})图片将删除，结果${deleteImage(pid)}")
+        sendMessage("作品(${pid})图片将删除，结果${EmptyArtWorkInfo.copy(pid = pid).delete()}")
     }
 
     @SubCommand
     @Description("删除指定用户作品")
-    suspend fun CommandSender.user(uid: Long) {
+    suspend fun CommandSender.user(uid: Long, record: Boolean = false) {
         val artworks = useMappers { it.artwork.userArtWork(uid) }
-        useMappers { mappers ->
+        if (record) useMappers { mappers ->
             mappers.artwork.deleteByUid(uid = uid, comment = "command delete user in ${OffsetDateTime.now()}")
         }
         sendMessage("USER(${uid})共${artworks.size}个作品需要删除")
         artworks.forEach {
             logger.info { "作品(${it.pid})信息将从缓存移除" }
-            deleteImage(it.pid)
+            it.delete()
         }
         sendMessage("删除完毕")
     }
@@ -53,21 +53,47 @@ object PixivDeleteCommand : CompositeCommand(
     private val ranges = (0 until OFFSET_MAX step OFFSET_STEP).map { offset -> offset until (offset + OFFSET_STEP) }
 
     @SubCommand
-    @Description("删除小于指定收藏数作品（用于处理漫画作品）")
-    suspend fun CommandSender.bookmarks(bookmarks: Long) {
+    @Description("删除小于指定收藏数作品")
+    suspend fun CommandSender.bookmarks(min: Long, record: Boolean = false) {
         ranges.forEach { interval ->
             logger.verbose { "开始检查[$interval]" }
             val artworks = useMappers { mappers ->
-                mappers.artwork.artworks(interval).filter { it.totalBookmarks < bookmarks }
+                mappers.artwork.artworks(interval).filter { it.totalBookmarks < min }
             }
-            sendMessage("{<$bookmarks}(${interval})共${artworks.size}个作品需要删除")
-            artworks.forEach {
-                logger.info { "作品(${it.pid})信息将从缓存移除" }
-                useMappers { mappers ->
-                    mappers.artwork.deleteByPid(
+            if (artworks.isEmpty()) return@forEach
+            sendMessage("{$min}(${interval})共${artworks.size}个作品需要删除")
+            useMappers { mappers ->
+                artworks.forEach {
+                    logger.info { "作品(${it.pid})信息将从缓存移除" }
+                    if (record) mappers.artwork.deleteByPid(
                         pid = it.pid,
-                        comment = "command delete bookmarks $bookmarks in ${OffsetDateTime.now()}"
+                        comment = "command delete bookmarks $min in ${OffsetDateTime.now()}"
                     )
+                    it.delete()
+                }
+            }
+            sendMessage("删除完毕")
+        }
+    }
+
+    @SubCommand
+    @Description("删除大于指定页数作品（用于处理漫画作品）")
+    suspend fun CommandSender.page(max: Int, record: Boolean = false) {
+        ranges.forEach { interval ->
+            logger.verbose { "开始检查[$interval]" }
+            val artworks = useMappers { mappers ->
+                mappers.artwork.artworks(interval).filter { it.pageCount > max }
+            }
+            if (artworks.isEmpty()) return@forEach
+            sendMessage("[$max](${interval})共${artworks.size}个作品需要删除")
+           useMappers { mappers ->
+                artworks.forEach {
+                    logger.info { "作品(${it.pid})信息将从缓存移除" }
+                    if (record) mappers.artwork.deleteByPid(
+                        pid = it.pid,
+                        comment = "command delete page_count $max in ${OffsetDateTime.now()}"
+                    )
+                    it.delete()
                 }
             }
             sendMessage("删除完毕")
